Comprehensive Implementation for Reader-Specific GPI/GPO Configuration

+++ OctaneTagJobControlAPI/Strategies/Base/Configuration/EnduranceTestConfiguration.cs
@@ -26,10 +26,10 @@
         public int SuccessCountLogIntervalSeconds { get; set; } = 5;
 
         public bool EnableGpiTrigger { get; set; } = false;
-        public int GpiPort { get; set; } = 1;
+        public ushort GpiPort { get; set; } = 1;
         public bool GpiTriggerState { get; set; } = true;
         public bool EnableGpoOutput { get; set; } = false;
-        public int GpoPort { get; set; } = 1;
+        public ushort GpoPort { get; set; } = 1;
         public int GpoVerificationTimeoutMs { get; set; } = 1000;
     }
 }
